Updated PageData.yaml and template.html. Also added a new template filter to the app.
Extracted content related to sidebar navigation and reources from HTML file into YAML.
Added comments to explain the structure of the YAML file for easier use.

New template filter will allow us to have equal width columns in tables if needed.
(Example use case for filter is for the table that displays the schedule of the event)